+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 This is a user-focused changelog. The entries here are things that affect gameplay, not technical details behind the scenes.
 
-* DEV: Added a new league setting "Number of Periods Per Game" to let you change from the default of 4 quarters to whatever you want.
+* DEV: Added a new league setting "Number of Periods Per Game" to let you change from the default of 4 quarters to whatever you want. In FBGM, if the number of periods is odd, there is no halftime. In BBGM, I also fixed the logic for which team gets possession at the start of a period (it was actually implemented like football previously, I have no idea what I was thinking when I did that).
 
 * 2021-01-11: If you live sim the last game of the playoffs, confetti drops after the game ends. [[1]](https://old.reddit.com/r/BasketballGM/comments/kviwek/if_you_live_sim_the_last_game_of_the_playoffs/)